% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotators.R
\name{word_annotator}
\alias{word_annotator}
\alias{person_annotator}
\alias{location_annotator}
\alias{organization_annotator}
\alias{date_annotator}
\alias{money_annotator}
\alias{percent_annotator}
\title{Annotators}
\usage{
word_annotator()

person_annotator()

location_annotator()

organization_annotator()

date_annotator()

money_annotator()

percent_annotator()
}
\value{
Returns an annotator for entities or words.
}
\description{
A wrapper for \code{\link[openNLP]{Maxent_Entity_Annotator}} and
\code{\link[openNLP]{Maxent_Word_Token_Annotator}}.
}
\seealso{
\code{\link[openNLP]{Maxent_Entity_Annotator}},
\code{\link[openNLP]{Maxent_Word_Token_Annotator}}
}
