% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entity-package.R
\docType{package}
\name{entity}
\alias{entity}
\alias{entity-package}
\alias{package-entity}
\title{Named Entity Extaction}
\description{
A wrapper for \pkg{NLP} and \pkg{openNLP} to facilitate named entity extraction.
}

