% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/named_entity.R
\name{named_entity}
\alias{named_entity}
\title{Named Entity Recognition}
\usage{
named_entity(text.var, entity.annotator, word.annotator = word_annotator(),
  element.chunks = floor(2000 * (23.5/mean(sapply(text.var, nchar), na.rm =
  TRUE))))
}
\arguments{
\item{text.var}{The text string variable.}

\item{entity.annotator}{An character vector identifying an entity recognition
annotator (\code{c("person_annotator", "location_annotator", "date_annotator",
"money_annotator", "percent_annotator")}.  See \code{?annotators}.}

\item{word.annotator}{A word annotator.}

\item{element.chunks}{The number of elements to include in a chunk. Chunks are
passed through an \code{\link[base]{lapply}} and size is kept within a tolerance
because of memory allocation in the tagging process with \pkg{Java}.}
}
\value{
Returns a data.frame of named entities and frequencies.
}
\description{
A wrapper for \pkg{NLP},/\pkg{openNLP}'s named entity recogniton annotation
tools.
}
\examples{
\dontrun{
data(presidential_debates_2012)

peoples <- named_entity(presidential_debates_2012$dialogue, 'person_annotator')
unlist(peoples)

orgs <-named_entity(presidential_debates_2012$dialogue, 'organization_annotator')
unlist(orgs)

dates <-named_entity(presidential_debates_2012$dialogue, 'date_annotator')
unlist(dates)

library(dplyr)
presidential_debates_2012$organizations <- named_entity(
    presidential_debates_2012$dialogue,
    'organization_annotator'
)

presidential_debates_2012 \%>\%
    {.[!sapply(.$organizations, is.null), ]} \%>\%
    rowwise() \%>\%
    mutate(organizations = paste(organizations, collapse=", ")) \%>\%
    select(person, time, organizations)

library(tidyr)
presidential_debates_2012 \%>\%
    {.[!sapply(.$organizations, is.null), ]} \%>\%
    unnest() \%>\%
    select(person, time, organizations)
}
}
\seealso{
\code{\link[openNLP]{Maxent_Entity_Annotator}}
}
\keyword{entity}
\keyword{named}
\keyword{ner}

