% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/named_entity.R
\name{plot.entity}
\alias{plot.entity}
\title{Plots a plot.entity Object}
\usage{
\method{plot}{entity}(x, min = 1, alphabetical = FALSE, ...)
}
\arguments{
\item{x}{An \code{entity} object.}

\item{min}{Minimum frequency of included entities.}

\item{alphabetical}{logical.  Should rows be arranged alphabetically by entity
or by frequency.}

\item{\ldots}{ignored.}
}
\description{
Plots a plot.entity object
}

