% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_entity.R
\name{print.entity}
\alias{print.entity}
\title{Prints a entity Object}
\usage{
\method{print}{entity}(x, ...)
}
\arguments{
\item{x}{An \code{entity} object.}

\item{\ldots}{ignored.}
}
\description{
Prints a entity object
}
