% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entity-package.R
\docType{data}
\name{wiki}
\alias{wiki}
\title{Bell Labs Wikipedia Article}
\format{A character vector with 7 elements}
\usage{
data(wiki)
}
\description{
A dataset containing a character vector of an excerpt from Wikipedia about
Bell Labs with an extra final sentence to include percent and money when
extracting entities.
}
\references{
\url{https://en.wikipedia.org/wiki/Bell_Labs}
}
\keyword{datasets}

