% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/person_entity.R
\name{person_entity}
\alias{person_entity}
\title{Named Person Recognition}
\usage{
person_entity(text.var, entity.annotator = "person_annotator",
  word.annotator = word_annotator(), element.chunks = floor(2000 *
  (23.5/mean(sapply(text.var, nchar), na.rm = TRUE))))
}
\arguments{
\item{text.var}{The text string variable.}

\item{entity.annotator}{An character vector identifying an entity recognition
annotator (\code{c("person_annotator", "location_annotator", "date_annotator",
"money_annotator", "percent_annotator")}.  See \code{?annotators}.}

\item{word.annotator}{A word annotator.}

\item{element.chunks}{The number of elements to include in a chunk. Chunks are
passed through an \code{\link[base]{lapply}} and size is kept within a tolerance
because of memory allocation in the tagging process with \pkg{Java}.}
}
\value{
Returns a data.frame of named entities and frequencies.
}
\description{
A wrapper for \pkg{NLP},/\pkg{openNLP}'s named person recogniton annotation.
}
\examples{
\dontrun{
data(presidential_debates_2012)

peoples <- person_entity(presidential_debates_2012$dialogue)
unlist(peoples)

library(dplyr)
presidential_debates_2012$persons <- person_entity(presidential_debates_2012$dialogue)

presidential_debates_2012 \%>\%
    {.[!sapply(.$persons, is.null), ]} \%>\%
    rowwise() \%>\%
    mutate(persons = paste(organizations, collapse=", ")) \%>\%
    select(person, time, persons)

library(tidyr)
presidential_debates_2012 \%>\%
    {.[!sapply(.$persons, is.null), ]} \%>\%
    unnest() \%>\%
    select(person, time, persons)
}
}
\keyword{people}
\keyword{person}

